pyproject.toml
[build-system]
requires = [ "poetry>=0.12", "setuptools", "mypy", "returns" ]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "synth-a-py"
version = "1.6.0"
description = "Project configuration as code"
authors = [ "Joseph Egan <joseph.s.egan@gmail.com>" ]
license = "MIT"
repository = "https://github.com/eganjs/synth-a-py"
readme = "README.md"

[tool.poetry.build]
script = "build.py"

  [tool.poetry.dependencies]
  python = "^3.6"
  returns = "^0.14.0"
  "ruamel.yaml" = "^0.16.12"
  toml = "^0.10.1"

    [tool.poetry.dependencies.contextvars]
    version = "^2.4"
    python = "~3.6"

  [tool.poetry.dev-dependencies]
  black = "^20.8b1"
  flake8 = "^3.8.4"
  isort = "^5.6"
  mypy = "^0.812"
  pyprojroot = "^0.2.0"
  pytest = "^6"
  toml = "^0.10.1"
  jedi = "^0.17.2"
  typing-extensions = "^3.7.4.3"
